Reorganize components to prepare for routing
